services/gateway/requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
motor>=3.3.0
python-dotenv>=1.0.0
pydantic>=2.5.0
httpx>=0.25.0
chromadb>=0.4.0
duckduckgo-search>=4.0.0
openai>=1.3.0
anthropic>=0.7.0
google-generativeai>=0.3.0
